Rename neon wrap to neon alias
