+++ contrib/standalone_buffer/Makefile
@@ -50,7 +50,7 @@ $(SETUPROOT)/BufferExpander.o: $(SETUPROOT)/modmesh/buffer/BufferExpander.cpp Ma
 $(SETUPROOT)/SimpleArray.o: $(SETUPROOT)/modmesh/buffer/SimpleArray.cpp Makefile
 	c++ $(CFLAGS) -c $< -o $@
 
-$(SETUPROOT)/neon_wrap.o: $(SETUPROOT)/modmesh/simd/neon/neon_wrap.cpp Makefile
+$(SETUPROOT)/neon_alias.o: $(SETUPROOT)/modmesh/simd/neon/neon_alias.cpp Makefile
 	c++ $(CFLAGS) -c $< -o $@
 
 $(SETUPROOT)/modbuf.o: modbuf.cpp
@@ -63,7 +63,7 @@ OBJS = \
 	$(SETUPROOT)/wrap_SimpleArray.o \
 	$(SETUPROOT)/wrap_SimpleArrayPlex.o \
 	$(SETUPROOT)/buffer_pymod.o \
-	$(SETUPROOT)/neon_wrap.o \
+	$(SETUPROOT)/neon_alias.o \
 	$(SETUPROOT)/modbuf.o
 
 $(MODPATH): Makefile $(OBJS)
